updated the install/uninstall scripts & added new test methods

+++ uninstall.sh
@@ -7,29 +7,44 @@ if [ "$EUID" -ne 0 ]
   exit
 fi
 
+# Ensuring system is debian based
+if [ "$OS_CHK" != "debian" ]; then
+    echo "Unfortunately this install script was written for debian based distributions only, sorry!"
+    exit
+fi
+
 # Removing the Sherlock project file and soft link
 rm /usr/bin/sherlock
 rm /opt/Sherlock -rf
 
-# Removing the jexboss dependency
-pip uninstall -r /opt/jexboss/requires.txt
-rm /opt/jexboss/ -rf
-
 # Removing the XSStrike dependency
 pip3 uninstall -r /opt/XSStrike/requirements.txt
 rm /opt/XSStrike/ -rf
-rm /usr/bin/xsstrike
 
 # Removing the SpiderLabs Nmap Script dependency
 rm /opt/SpiderLabs/Nmap-Tools -rf
 rm /usr/share/nmap/scripts/http-screenshot.nse
 
-# Downloading the Vulners Nmap Script
+# Removing the Vulners Nmap Script dependency
 rm /opt/vulnersCom/nmap-vulners -rf
 rm /usr/share/nmap/scripts/nmap-vulners/vulners.nse
 
+# Removing ssh-audit dependency
+rm -rf /opt/ssh-audit
+
+# Removing the sshscan dependency
+rm -rf /opt/SSHScan/
+
+# Removing npm, nodejs, and retirejs dependency
+npm uninstall retire -g
+npm uninstall npm -g
+
+# Removing the SpiderLabs Nmap Script dependency
+rm -rf /opt/nmap-vulners
+rm -rf /usr/share/nmap/scripts/vulners.nse
+
 # Removing remaining dependencies
-apt remove halberd sublist3r theharvester metagoofil nikto dirb nmap sn1pe masscan arachni sslscan testssl jexboss grabber golismero docker -y
+apt remove halberd sublist3r theharvester metagoofil nikto nmap dnsrecon python3 masscan arachni sslscan testssl golismero gobuster docker-ce docker-ce-cli containerd.io nodejs -y
 pip uninstall halberd
 
 # Done!
